ai_service/facebook_scraper/requirements.txt
selenium==4.12.0
webdriver-manager==4.0.1
beautifulsoup4>=4.12.2
mysql-connector-python>=8.1.0
python-dotenv>=1.0.0
psutil>=5.9.5
flask>=2.3.3
requests>=2.31.0
gunicorn>=21.2.0 